WIP factoring out asset functionality into it's own methods
