+++ tools/dfx-orbit/src/cli/asset/evidence.rs
@@ -36,4 +36,6 @@ impl DfxOrbit {
         let canister_agent = self.canister_agent(canister_id)?;
         Ok(ic_asset::compute_evidence(&canister_agent, sources, &self.logger).await?)
     }
+
+    // TODO: check_evidence function
 }
